==> Jupyter/Step_2C_PreprocessData_Gestures.ipynb <==
# First approach to gesture recognition
This notebook creates one image per gesture by simply summing up all images during the gesture. This ways, we tried to recognize gesures with a simple CNN.

%matplotlib inline

from scipy.odr import *
from scipy.stats import *
import numpy as np
import pandas as pd
import os
import time
import matplotlib.pyplot as plt
import ast
from multiprocessing import Pool

import scipy

from IPython import display
from matplotlib.patches import Rectangle

from sklearn.metrics import mean_squared_error
import json

import scipy.stats as st
from sklearn.metrics import r2_score


from matplotlib import cm
from mpl_toolkits.mplot3d import axes3d
import matplotlib.pyplot as plt

import copy

from sklearn.model_selection import LeaveOneOut, LeavePOut

from multiprocessing import Pool, cpu_count
import cv2

dfAll = pd.read_pickle("PklData/AllData.pkl")
dfAll.head()

df_actual = dfAll[(dfAll.Actual_Data == True) & (dfAll.Is_Pause == False)]
df_actual.head()

print("all: %s, actual data: %s" % (len(dfAll), len(df_actual)))

%%time
def is_max(df):
    df_temp = df.copy(deep=True)
    max_version = df_temp.RepetitionID.max()
    df_temp["IsMax"] = np.where(df_temp.RepetitionID == max_version, True, False)
    df_temp["MaxRepetition"] = [max_version] * len(df_temp)
    return df_temp

df_filtered = df_actual.copy(deep=True)
df_grp = df_filtered.groupby([df_filtered.userID, df_filtered.TaskID, df_filtered.VersionID])
pool = Pool(cpu_count() - 1)
result_lst = pool.map(is_max, [grp for name, grp in df_grp])
df_filtered = pd.concat(result_lst)
df_filtered = df_filtered[df_filtered.IsMax == True]
pool.close()
print("actual: %s, filtered data: %s" % (len(df_actual), len(df_filtered)))

df_filtered.to_pickle("PklData/dfFiltered.pkl")

df_filtered = pd.read_pickle("PklData/dfFiltered.pkl")

print("actual: %s, filtered data: %s" % (len(df_actual), len(df_filtered)))

def detect_blobs(index, row):
    image = row.Image
    image = image.reshape(27, 15)
    temp, thresh = cv2.threshold(cv2.bitwise_not(image), 200, 255, cv2.THRESH_BINARY)
    contours, hierarchy = cv2.findContours(thresh, cv2.RETR_TREE, cv2.CHAIN_APPROX_SIMPLE)
    contours = [a for a in contours if cv2.contourArea(a) > 5 and cv2.contourArea(a) < 255]
    if len(contours) > 0:
        if (row.TaskID not in [6, 7, 23, 24]):
            # only one blob
            contours.sort(key=lambda a: cv2.contourArea(a))
            max_contour = contours[-1]
            xmax, ymax = np.max(max_contour.reshape(len(max_contour),2), axis=0)
            xmin, ymin = np.min(max_contour.reshape(len(max_contour),2), axis=0)
            croped_im = np.zeros((27,15))
            blob = image[max(ymin - 1, 0):min(ymax + 1, image.shape[0]),max(xmin - 1, 0):min(xmax + 1, image.shape[1])]
            croped_im[0:blob.shape[0],0:blob.shape[1]] = blob
            return (1, [croped_im])
        else:
            # two blobs
            blobs = []
            for i in range(1, 2):
                contours.sort(key=lambda a: cv2.contourArea(a))
                max_contour = contours[-1 * i]
                xmax, ymax = np.max(max_contour.reshape(len(max_contour),2), axis=0)
                xmin, ymin = np.min(max_contour.reshape(len(max_contour),2), axis=0)
                croped_im = np.zeros((27,15))
                blob = image[max(ymin - 1, 0):min(ymax + 1, image.shape[0]),max(xmin - 1, 0):min(xmax + 1, image.shape[1])]
                croped_im[0:blob.shape[0],0:blob.shape[1]] = blob
                blobs.append(blobs)
            return (2, blobs)
    else:
        return (0, np.zeros((27,15)))

%%time
pool = Pool(cpu_count())
df_croped = df_filtered.copy()
temp_blobs = pool.starmap(detect_blobs, df_filtered.iterrows())
df_croped["BlobCount"] = [a[0] for a in temp_blobs]
df_croped["Blobs"] = [a[1] for a in temp_blobs]
df_croped.head()

df_blobs = df_croped[df_croped.BlobCount >= 1]
print("detected blobs: %s" % len(df_blobs))
df_blobs.head()

df_blobs.to_pickle("PklData/df_blobs_gestures.pkl")

# Build Gestures

def detect_and_return_blobs(row):
    image = row[1].Image
    image = image.reshape(27, 15)
    temp, thresh = cv2.threshold(cv2.bitwise_not(image), 200, 255, cv2.THRESH_BINARY)
    contours, hierarchy = cv2.findContours(thresh, cv2.RETR_TREE, cv2.CHAIN_APPROX_SIMPLE)
    contours = [a for a in contours if cv2.contourArea(a) > 5 and cv2.contourArea(a) < 255]
    if len(contours) > 0:
        if (row[1].TaskID not in [6, 7, 23, 24]):
            # only one blob
            contours.sort(key=lambda a: cv2.contourArea(a))
            max_contour = contours[-1]
            xmax, ymax = np.max(max_contour.reshape(len(max_contour),2), axis=0)
            xmin, ymin = np.min(max_contour.reshape(len(max_contour),2), axis=0)
            croped_im = np.zeros((27,15))
            blob = image[max(ymin - 1, 0):min(ymax + 1, image.shape[0]),max(xmin - 1, 0):min(xmax + 1, image.shape[1])]
            croped_im[max(ymin - 1, 0):min(ymax + 1, image.shape[0]),max(xmin - 1, 0):min(xmax + 1, image.shape[1])] = blob
            return [croped_im]
        else:
            # two blobs
            blobs = []
            for i in range(1, 2):
                contours.sort(key=lambda a: cv2.contourArea(a))
                max_contour = contours[-1 * i]
                xmax, ymax = np.max(max_contour.reshape(len(max_contour),2), axis=0)
                xmin, ymin = np.min(max_contour.reshape(len(max_contour),2), axis=0)
                croped_im = np.zeros((27,15))
                blob = image[max(ymin - 1, 0):min(ymax + 1, image.shape[0]),max(xmin - 1, 0):min(xmax + 1, image.shape[1])]
                croped_im[max(ymin - 1, 0):min(ymax + 1, image.shape[0]),max(xmin - 1, 0):min(xmax + 1, image.shape[1])] = blob
                blobs.append(croped_im)
            return blobs
    else:
        return np.zeros((27,15))

%%time
data_points = []

for UserID in df_blobs.userID.unique():
    print(UserID)
    for TaskID in df_blobs[df_blobs.userID == UserID].TaskID.unique():
        for VersionID in df_blobs[(df_blobs.userID == UserID) & (df_blobs.TaskID == TaskID)].VersionID.unique():
            for RepetitionID in df_blobs[(df_blobs.userID == UserID) & (df_blobs.TaskID == TaskID) & (df_blobs.VersionID == VersionID)].RepetitionID.unique():
                data = df_blobs[(df_blobs.userID == UserID) & (df_blobs.TaskID == TaskID) & (df_blobs.VersionID == VersionID) & (df_blobs.RepetitionID == RepetitionID)]
                gesture_im = np.zeros((27,15))
                for row in data.iterrows():
                    blobs = detect_and_return_blobs(row)
                    for blob in blobs:
                        gesture_im = gesture_im + blob
                #gesture_im = np.clip(gesture_im, 0, 255)
                data_points.append([UserID, TaskID, VersionID, RepetitionID, gesture_im])

print(len(data_points))

df_gestures = pd.DataFrame(data_points)
df_gestures.columns = ["userID", "TaskID", "VersionID", "RepetitionID", "Gesture"]
df_gestures.head()

df_gestures.to_pickle("PklData/df_gestures.pkl")

# display Gesture

user_id = 4
len(df_gestures[(df_gestures.userID == user_id) & (df_gestures.TaskID == 15)])


user_id = 4
task_id = 15
plt.clf()
for k in range(0, len(df_gestures[(df_gestures.userID == user_id) & (df_gestures.TaskID == task_id)])):
    data = df_gestures[(df_gestures.userID == user_id) & (df_gestures.TaskID == task_id)].Gesture.iloc[k]
    plt.figure(figsize=(6, 6))
    ax = plt.gca()
    plt.imshow(data, cmap='gray', vmin=0, vmax=255)
    # Loop over data dimensions and create text annotations.
    for i in range(0, data.shape[0]):
        for j in range(0, data.shape[1]):
            text = ax.text(j, i, int(data[i, j]),
                           ha="center", va="center", color="cyan", fontsize=1)
    #plt.savefig("./GestureImages/gesture_%s_%s_%s.png" % (task_id, user_id, k), bbox_inches='tight')
    plt.show()
